Added a speedometer mod.
