+++ SuperflightMods.sln
@@ -9,7 +9,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnlimitedFlight", "Unlimite
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RunStatistics", "RunStatistics\RunStatistics.csproj", "{4E8218C2-1A37-49FB-9C36-44E33AEE2716}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TimeElapsed", "TimeElapsed\TimeElapsed.csproj", "{B9D902FB-16CA-40D0-855F-F26CA77E6180}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Speedometer", "Speedometer\Speedometer.csproj", "{B9D902FB-16CA-40D0-855F-F26CA77E6180}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NextWorld", "NextWorld\NextWorld.csproj", "{6F43B2B1-0FF5-4CE3-A834-749638BF2407}"
 EndProject
